BS-11: removed user database functions from model
